refactor: init v3
