+++ index.html
@@ -8,6 +8,18 @@
   </head>
   <body>
     NoneOS
-    <script src="import-proxy.js"></script>
+    <script>
+      if ("serviceWorker" in navigator) {
+        navigator.serviceWorker
+          .register("/proxy.js", { type: "module" })
+          .then((reg) => {
+            setTimeout(() => {
+              // 一小时后尝试自动更新
+              reg.update();
+            }, 60 * 60 * 1000);
+          })
+          .catch((err) => {});
+      }
+    </script>
   </body>
 </html>
