+++ olds/v3/test/fs/fs.html
@@ -4,10 +4,14 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>fs test</title>
     <script src="https://cdn.jsdelivr.net/gh/kirakiray/ofa.js/test/ok.js"></script>
-    <title>file system test</title>
   </head>
   <body>
-    <script src="base-test.mjs" type="module"></script>
+    <script src="./base-test.js" type="module"></script>
+    <script src="./remove-test.js" type="module"></script>
+    <script src="./proxy-test.js" type="module"></script>
+    <script src="./getall-test.js" type="module"></script>
+    <script src="./normal-test.js" type="module"></script>
   </body>
 </html>
