+++ olds/v3/test/fs/proxy-test.js
