+++ olds/v3/test/fs/remove-test.js
