+++ system/fs/mkdir.mjs
@@ -1,0 +1,5 @@
+import { writeDB, readDB, initRoot } from "./base.mjs";
+
+export default function mkdir() {
+    
+}
